+++ packages/react/src/errorboundary.tsx
@@ -4,6 +4,8 @@ import { isError, logger } from '@sentry/utils';
 import hoistNonReactStatics from 'hoist-non-react-statics';
 import * as React from 'react';
 
+import { DEBUG_BUILD } from './debug-build';
+
 export function isAtLeastReact17(version: string): boolean {
   const major = version.match(/^([^.]+)/);
   return major !== null && parseInt(major[0]) >= 17;
@@ -21,12 +23,12 @@ export type FallbackRender = (errorData: {
 export type ErrorBoundaryProps = {
   children?: React.ReactNode | (() => React.ReactNode);
   /** If a Sentry report dialog should be rendered on error */
-  showDialog?: boolean;
+  showDialog?: boolean | undefined;
   /**
    * Options to be passed into the Sentry report dialog.
    * No-op if {@link showDialog} is false.
    */
-  dialogOptions?: ReportDialogOptions;
+  dialogOptions?: ReportDialogOptions | undefined;
   /**
    * A fallback component that gets rendered when the error boundary encounters an error.
    *
@@ -35,17 +37,17 @@ export type ErrorBoundaryProps = {
    * the error, the component stack, and an function that resets the error boundary on error.
    *
    */
-  fallback?: React.ReactElement | FallbackRender;
+  fallback?: React.ReactElement | FallbackRender | undefined;
   /** Called when the error boundary encounters an error */
-  onError?(error: Error, componentStack: string, eventId: string): void;
+  onError?: ((error: Error, componentStack: string, eventId: string) => void) | undefined;
   /** Called on componentDidMount() */
-  onMount?(): void;
+  onMount?: (() => void) | undefined;
   /** Called if resetError() is called from the fallback render props function  */
-  onReset?(error: Error | null, componentStack: string | null, eventId: string | null): void;
+  onReset?: ((error: Error | null, componentStack: string | null, eventId: string | null) => void) | undefined;
   /** Called on componentWillUnmount() */
-  onUnmount?(error: Error | null, componentStack: string | null, eventId: string | null): void;
+  onUnmount?: ((error: Error | null, componentStack: string | null, eventId: string | null) => void) | undefined;
   /** Called before the error is captured by Sentry, allows for you to add tags or context using the scope */
-  beforeCapture?(scope: Scope, error: Error | null, componentStack: string | null): void;
+  beforeCapture?: ((scope: Scope, error: Error | null, componentStack: string | null) => void) | undefined;
 };
 
 type ErrorBoundaryState =
@@ -208,7 +210,7 @@ class ErrorBoundary extends React.Component<ErrorBoundaryProps, ErrorBoundarySta
       }
 
       if (fallback) {
-        __DEBUG_BUILD__ && logger.warn('fallback did not produce a valid ReactElement');
+        DEBUG_BUILD && logger.warn('fallback did not produce a valid ReactElement');
       }
 
       // Fail gracefully if no fallback provided or is not valid