+++ packages/replay/src/coreHandlers/util/networkUtils.ts
@@ -2,6 +2,7 @@ import type { TextEncoderInternal } from '@sentry/types';
 import { dropUndefinedKeys, logger, stringMatchesSomePattern } from '@sentry/utils';
 
 import { NETWORK_BODY_MAX_SIZE, WINDOW } from '../../constants';
+import { DEBUG_BUILD } from '../../debug-build';
 import type {
   NetworkBody,
   NetworkMetaWarning,
@@ -75,11 +76,11 @@ export function getBodyString(body: unknown): [string | undefined, NetworkMetaWa
       return [_serializeFormData(body)];
     }
   } catch {
-    __DEBUG_BUILD__ && logger.warn('[Replay] Failed to serialize body', body);
+    DEBUG_BUILD && logger.warn('[Replay] Failed to serialize body', body);
     return [undefined, 'BODY_PARSE_ERROR'];
   }
 
-  __DEBUG_BUILD__ && logger.info('[Replay] Skipping network body because of body type', body);
+  DEBUG_BUILD && logger.info('[Replay] Skipping network body because of body type', body);
 
   return [undefined];
 }