+++ packages/replay/src/replay.ts
@@ -15,6 +15,7 @@ import {
 import { ClickDetector } from './coreHandlers/handleClick';
 import { handleKeyboardEvent } from './coreHandlers/handleKeyboardEvent';
 import { setupPerformanceObserver } from './coreHandlers/performanceObserver';
+import { DEBUG_BUILD } from './debug-build';
 import { createEventBuffer } from './eventBuffer';
 import { clearSession } from './session/clearSession';
 import { loadOrCreateSession } from './session/loadOrCreateSession';
@@ -726,9 +727,9 @@ export class ReplayContainer implements ReplayContainerInterface {
 
   /** A wrapper to conditionally capture exceptions. */
   private _handleException(error: unknown): void {
-    __DEBUG_BUILD__ && logger.error('[Replay]', error);
+    DEBUG_BUILD && logger.error('[Replay]', error);
 
-    if (__DEBUG_BUILD__ && this._options._experiments && this._options._experiments.captureExceptions) {
+    if (DEBUG_BUILD && this._options._experiments && this._options._experiments.captureExceptions) {
       captureException(error);
     }
   }
@@ -1045,7 +1046,7 @@ export class ReplayContainer implements ReplayContainerInterface {
     const replayId = this.getSessionId();
 
     if (!this.session || !this.eventBuffer || !replayId) {
-      __DEBUG_BUILD__ && logger.error('[Replay] No session or eventBuffer found to flush.');
+      DEBUG_BUILD && logger.error('[Replay] No session or eventBuffer found to flush.');
       return;
     }
 
@@ -1135,7 +1136,7 @@ export class ReplayContainer implements ReplayContainerInterface {
     }
 
     if (!this.checkAndHandleExpiredSession()) {
-      __DEBUG_BUILD__ && logger.error('[Replay] Attempting to finish replay event after session expired.');
+      DEBUG_BUILD && logger.error('[Replay] Attempting to finish replay event after session expired.');
       return;
     }
 
@@ -1193,7 +1194,7 @@ export class ReplayContainer implements ReplayContainerInterface {
     try {
       await this._flushLock;
     } catch (err) {
-      __DEBUG_BUILD__ && logger.error(err);
+      DEBUG_BUILD && logger.error(err);
     } finally {
       this._debouncedFlush();
     }