+++ packages/replay/src/util/log.ts
@@ -1,11 +1,13 @@
 import { getCurrentHub } from '@sentry/core';
 import { logger } from '@sentry/utils';
 
+import { DEBUG_BUILD } from '../debug-build';
+
 /**
  * Log a message in debug mode, and add a breadcrumb when _experiment.traceInternals is enabled.
  */
 export function logInfo(message: string, shouldAddBreadcrumb?: boolean): void {
-  if (!__DEBUG_BUILD__) {
+  if (!DEBUG_BUILD) {
     return;
   }
 
@@ -21,7 +23,7 @@ export function logInfo(message: string, shouldAddBreadcrumb?: boolean): void {
  * This is necessary when the breadcrumb may be added before the replay is initialized.
  */
 export function logInfoNextTick(message: string, shouldAddBreadcrumb?: boolean): void {
-  if (!__DEBUG_BUILD__) {
+  if (!DEBUG_BUILD) {
     return;
   }