+++ packages/tracing-internal/src/browser/metrics/index.ts
@@ -4,6 +4,7 @@ import { getActiveTransaction } from '@sentry/core';
 import type { Measurements } from '@sentry/types';
 import { browserPerformanceTimeOrigin, htmlTreeAsString, logger } from '@sentry/utils';
 
+import { DEBUG_BUILD } from '../../common/debug-build';
 import {
   addClsInstrumentationHandler,
   addFidInstrumentationHandler,
@@ -121,7 +122,7 @@ function _trackCLS(): () => void {
       return;
     }
 
-    __DEBUG_BUILD__ && logger.log('[Measurements] Adding CLS');
+    DEBUG_BUILD && logger.log('[Measurements] Adding CLS');
     _measurements['cls'] = { value: metric.value, unit: '' };
     _clsEntry = entry as LayoutShift;
   });
@@ -135,7 +136,7 @@ function _trackLCP(): () => void {
       return;
     }
 
-    __DEBUG_BUILD__ && logger.log('[Measurements] Adding LCP');
+    DEBUG_BUILD && logger.log('[Measurements] Adding LCP');
     _measurements['lcp'] = { value: metric.value, unit: 'millisecond' };
     _lcpEntry = entry as LargestContentfulPaint;
   });
@@ -151,7 +152,7 @@ function _trackFID(): () => void {
 
     const timeOrigin = msToSec(browserPerformanceTimeOrigin as number);
     const startTime = msToSec(entry.startTime);
-    __DEBUG_BUILD__ && logger.log('[Measurements] Adding FID');
+    DEBUG_BUILD && logger.log('[Measurements] Adding FID');
     _measurements['fid'] = { value: metric.value, unit: 'millisecond' };
     _measurements['mark.fid'] = { value: timeOrigin + startTime, unit: 'second' };
   });
@@ -165,7 +166,7 @@ export function addPerformanceEntries(transaction: Transaction): void {
     return;
   }
 
-  __DEBUG_BUILD__ && logger.log('[Tracing] Adding & adjusting spans using Performance API');
+  DEBUG_BUILD && logger.log('[Tracing] Adding & adjusting spans using Performance API');
   const timeOrigin = msToSec(browserPerformanceTimeOrigin);
 
   const performanceEntries = performance.getEntries();
@@ -200,11 +201,11 @@ export function addPerformanceEntries(transaction: Transaction): void {
         const shouldRecord = entry.startTime < firstHidden.firstHiddenTime;
 
         if (entry.name === 'first-paint' && shouldRecord) {
-          __DEBUG_BUILD__ && logger.log('[Measurements] Adding FP');
+          DEBUG_BUILD && logger.log('[Measurements] Adding FP');
           _measurements['fp'] = { value: entry.startTime, unit: 'millisecond' };
         }
         if (entry.name === 'first-contentful-paint' && shouldRecord) {
-          __DEBUG_BUILD__ && logger.log('[Measurements] Adding FCP');
+          DEBUG_BUILD && logger.log('[Measurements] Adding FCP');
           _measurements['fcp'] = { value: entry.startTime, unit: 'millisecond' };
         }
         break;
@@ -228,7 +229,7 @@ export function addPerformanceEntries(transaction: Transaction): void {
     // Generate TTFB (Time to First Byte), which measured as the time between the beginning of the transaction and the
     // start of the response in milliseconds
     if (typeof responseStartTimestamp === 'number') {
-      __DEBUG_BUILD__ && logger.log('[Measurements] Adding TTFB');
+      DEBUG_BUILD && logger.log('[Measurements] Adding TTFB');
       _measurements['ttfb'] = {
         value: (responseStartTimestamp - transaction.startTimestamp) * 1000,
         unit: 'millisecond',
@@ -258,8 +259,7 @@ export function addPerformanceEntries(transaction: Transaction): void {
       const normalizedValue = Math.abs((measurementTimestamp - transaction.startTimestamp) * 1000);
       const delta = normalizedValue - oldValue;
 
-      __DEBUG_BUILD__ &&
-        logger.log(`[Measurements] Normalized ${name} from ${oldValue} to ${normalizedValue} (${delta})`);
+      DEBUG_BUILD && logger.log(`[Measurements] Normalized ${name} from ${oldValue} to ${normalizedValue} (${delta})`);
       _measurements[name].value = normalizedValue;
     });
 
@@ -461,7 +461,7 @@ function _trackNavigator(transaction: Transaction): void {
 /** Add LCP / CLS data to transaction to allow debugging */
 function _tagMetricInfo(transaction: Transaction): void {
   if (_lcpEntry) {
-    __DEBUG_BUILD__ && logger.log('[Measurements] Adding LCP Data');
+    DEBUG_BUILD && logger.log('[Measurements] Adding LCP Data');
 
     // Capture Properties of the LCP element that contributes to the LCP.
 
@@ -483,7 +483,7 @@ function _tagMetricInfo(transaction: Transaction): void {
 
   // See: https://developer.mozilla.org/en-US/docs/Web/API/LayoutShift
   if (_clsEntry && _clsEntry.sources) {
-    __DEBUG_BUILD__ && logger.log('[Measurements] Adding CLS Data');
+    DEBUG_BUILD && logger.log('[Measurements] Adding CLS Data');
     _clsEntry.sources.forEach((source, index) =>
       transaction.setTag(`cls.source.${index + 1}`, htmlTreeAsString(source.node)),
     );